% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_axis_layout.r
\encoding{UTF-8}
\name{f_add_axis_layout}
\alias{f_add_axis_layout}
\title{Adding the main plotly graph layout}
\usage{
f_add_axis_layout(
  ply_1,
  s_data_type,
  df_click_legend,
  o_label_text,
  df_min_max,
  o_parameter,
  v_bg_grid_color,
  l_camera = NULL,
  s_option = "layout"
)
}
\arguments{
\item{ply_1}{is the plotly object created with the "build_graph" R script.}

\item{s_data_type}{is the data type: "normal" or "ir".}

\item{df_click_legend}{are legend item data information (name and status).}

\item{o_label_text}{is a reactive value including data information on the graph
label edition.}

\item{df_min_max}{is a data.frame including min-max values of each item
(\code{df_click_legend}) added in the main graph.}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels}

\item{v_bg_grid_color}{is a vector including the background/grid color.}

\item{l_camera}{is a list given the camera position of plot 3d (data type:
normal). The camera position is saved by the reactive value \code{o_zoom$coord}.}

\item{s_option}{has two values: "layout", "relayout". "layout" is used when the
display button is clicked and "relayout" is used when legend items are clicked.}
}
\description{
\code{f_add_axis_layout} returns a plotly object including (x, y, z) axis layout. This
function only concerns 2D/3D plot of the normal/ir data type.
}
