% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_centroid.r
\encoding{UTF-8}
\name{f_add_centroid}
\alias{f_add_centroid}
\title{Adding centroids in the main plotly graph}
\usage{
f_add_centroid(ply_1, df_click_legend, o_plot, o_parameter, l_graph_opt)
}
\arguments{
\item{ply_1}{is the plotly object created with the "build_graph" R script.}

\item{df_click_legend}{are legend item data information (name and status).}

\item{o_plot}{is a reactive value including \code{ply_1} data information.}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{l_graph_opt}{is the list of graph option (color, opacity, point type/size)
created by the \code{f_create_graph_opt_vector} function.}
}
\description{
\code{f_add_centroid} returns a plotly object including centroids. This function only
concerns 2D/3D plot of the normal data type (disabled if a calibration model is
applied).
}
