% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_conf_ellipsoid.r
\encoding{UTF-8}
\name{f_add_conf_ellipsoid}
\alias{f_add_conf_ellipsoid}
\title{Adding 95\% confidence ellipsoids in the main plotly graph}
\usage{
f_add_conf_ellipsoid(
  ply_1,
  v_group = "all",
  df_click_legend,
  o_plot,
  o_parameter,
  l_graph_opt
)
}
\arguments{
\item{ply_1}{is the plotly object created with the "build_graph" R script.}

\item{v_group}{is the Group variable levels used to applied the corresponding
method. If no Group variable is selected, then \code{v_group} = "all".}

\item{df_click_legend}{are legend item data information (name and status).}

\item{o_plot}{is a reactive value including \code{ply_1} data information.}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{l_graph_opt}{is the list of graph option (color, opacity, point type/size,
sorting) created by the \code{f_create_graph_opt_vector} function.}
}
\description{
\code{f_add_conf_ellipsoid} returns a plotly object including 95\% confidence
ellipsoids. This function only 2D plot of the normal data type (disabled if a
calibration model is applied).
}
