% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flag.r
\encoding{UTF-8}
\name{f_add_flag}
\alias{f_add_flag}
\title{Adding flag data in the main plotly graph}
\usage{
f_add_flag(
  s_data_type = "normal",
  ply_1,
  v_cond = rep(0, 3),
  df_click_legend,
  o_plot,
  o_parameter,
  df_previous_flag = NULL,
  e_current_flag
)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{ply_1}{is the plotly object created with the "build_graph" R script.}

\item{v_cond}{is a vector with binary values corresponding to 3 elements (flag,
qc1 and qc2). A zero value means no (qc1/qc2) flag data exist. If flag = 1, then
qc1 = 1 and/or qc2 = 1 (only temporal/ir data type).}

\item{df_click_legend}{are legend item data information (name and status).}

\item{o_plot}{is a reactive value including \code{ply_1} data information.}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{df_previous_flag}{are flag data saved in the e_data environment (first
category). This input is only used for the ir data type (default: NULL).}

\item{e_current_flag}{is the environment corresponding to flag data of the second
category.}
}
\description{
\code{f_add_flag} returns a plotly object corresponding to flag data information. Flag
data are separated into two categories:
\cr(1) Previous flags are data saved in the e_data environment (e_data$flag) and
splitted by qc value in the o_plot reactive value (o_plot$data_qc1,
o_plot$data_qc2);
\cr(2) Current flags are data saved in the e_current_flag environment.
}
