% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mean_spect.r
\encoding{UTF-8}
\name{f_add_mean_spect}
\alias{f_add_mean_spect}
\title{Adding mean spectra in the main plotly graph}
\usage{
f_add_mean_spect(ply_1, df_click_legend, o_plot, o_parameter)
}
\arguments{
\item{ply_1}{is the plotly object created with the "build_graph" R script.}

\item{df_click_legend}{are legend item data information (name and status).}

\item{o_plot}{is a reactive value including \code{ply_1} data information.}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}
}
\description{
\code{f_add_mean_spect} returns a plotly object including mean spectra. This function
only concerns the ir data type.
}
