% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph.r
\encoding{UTF-8}
\name{f_build_graph}
\alias{f_build_graph}
\title{Building the main plotly graph}
\usage{
f_build_graph(
  s_data_type = "normal",
  s_source = "graphic",
  v_dimension,
  o_click_button,
  df_click_legend,
  o_plot,
  o_parameter,
  l_axis_layout = NULL,
  o_picture_info,
  o_label_text,
  l_graph_opt = NULL
)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{s_source}{is the main panel (graph area) source.}

\item{v_dimension}{is the vector of the main panel dimension (width, height).}

\item{o_click_button}{is a reactive value including binary values related to
several buttons: clicked (1) or not (0).}

\item{df_click_legend}{are legend item data information (name and status).}

\item{o_plot}{is a reactive value including \code{ply_1} data information.}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{l_axis_layout}{is the list including camera eye position for 3D plot or (X,
Y) range for other plot types (used only for 2D plot and histplot).}

\item{o_picture_info}{is a reactive value including data information on the graph
picture.}

\item{o_label_text}{is a reactive value including data information on the graph
label edition.}

\item{l_graph_opt}{is the list of graph option (color, opacity, point type/size)
created by the \code{f_create_graph_opt_vector} function.}
}
\description{
\code{f_build_graph} returns the main plotly graph for all data types.
}
