% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_legend_items.r
\encoding{UTF-8}
\name{f_build_legend_items}
\alias{f_build_legend_items}
\title{Building legend item data of the main plotly graph}
\usage{
f_build_legend_items(
  s_data_type = "normal",
  i_proc_num = 1,
  o_parameter,
  o_cond,
  o_plot,
  o_stat_method = NULL,
  df_click_legend = data.frame()
)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{i_proc_num}{is the process number (2 values: 1, 2).}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{o_cond}{is a reactive value including binary values used as conditions in
the server (related to action buttons).}

\item{o_plot}{is a reactive value including data information corresponding to the
main plotly graph.}

\item{o_stat_method}{is a reactive value including data information on statistic
methods added to the main plotly graph.}

\item{df_click_legend}{are legend item data information (name and status).}
}
\description{
\code{f_build_legend_items} is used to build legend item data and update the
o_stat_method reactive value (created in the "WIDEa_server" R script) associated
to statistical methods (Statistics tab). Legend item data are saved in a two
columns data frame (item name and status: selected/unselected on the main plotly
graph). This data frame is returned for all data type. The function is separated
into 2 processes:
\cr(1) The first process is executed when a graph is created in the main panel by
clicking on the display button (left panel);
\cr(2) The second process is executed when the current graph is updated by
(de)selecting a statistical method from the Statistics tab (top panel) and
changing the graph when a calibration model is added (normal data type).
}
