% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_modal_dialog.r
\encoding{UTF-8}
\name{f_build_modal_dialog_expand}
\alias{f_build_modal_dialog_expand}
\alias{f_build_modal_dialog_data_option}
\alias{f_create_cond_data}
\alias{f_check_cond_data}
\alias{f_create_modal_code_var_list}
\alias{f_create_option_data}
\alias{f_update_option_data}
\alias{f_create_point_type_plotly}
\alias{f_create_bg_grid_data}
\alias{f_create_xy_cell_data}
\alias{f_create_xy_cell_plotly}
\title{Functions associated to a modal dialog box/popup window}
\usage{
f_build_modal_dialog_expand(
  s_title = NULL,
  s_size = "l",
  s_id,
  i_window_num = 1,
  b_info_var = F,
  b_code_var = F,
  b_info_cond = F
)

f_build_modal_dialog_data_option(i_window_num = 1)

f_create_cond_data(o_sdata_cond, s_selection = "multiple")

f_check_cond_data(o_sdata_cond, l_new_value)

f_create_modal_code_var_list(s_data_type, s_model, i_window_num, l_var)

f_create_option_data(
  s_option = "label",
  s_sub_option = NULL,
  o_option_value,
  s_data_type = NULL,
  s_plot_type = NULL,
  b_group = F,
  i_display = 0,
  b_pal_auto = F
)

f_update_option_data(
  s_option = "label",
  s_sub_option = NULL,
  s_exec = "clear",
  o_option_value,
  v_row = NULL,
  s_custom_text = NULL,
  s_data_type = NULL,
  s_plot_type = NULL,
  b_group = F,
  i_display = 0
)

f_create_point_type_plotly(s_dim_num = "2d")

f_create_bg_grid_data(o_bg_grid)

f_create_xy_cell_data(df_all, o_parameter, s_x_cell_var, s_y_cell_var)

f_create_xy_cell_plotly(df_xy_cell, i_graph_num)
}
\arguments{
\item{s_title}{is the tittle of the modal dialog window (NULL as default value).}

\item{s_size}{is the size of the modal dialog window. The size is a string value
and three values are avaible: "l" (large), "m" (medium) and "s" (small).}

\item{s_id}{is the UI's input ID.}

\item{i_window_num}{is an integer corresponding to the window number. Five values
are available: 1 (sub-data condition formula), 2 (f(x)), 3 (g(y)), 4 (h(z)) and 5
(weighted residuals function).}

\item{b_info_var}{is a boolean value used to select a datatable including a
warning message on missing variable(s).}

\item{b_code_var}{is a boolean value used to select a datatable including the
code/variable columns. Variables in the datatable are listed in \code{l_var} and are
associated to an unique code used to write the following function: f(x), g(y),
h(z), weighted residuals.}

\item{b_info_cond}{is a boolean value used to select a datatable including
sub-data condition information.}

\item{o_sdata_cond}{is a reactive value including all conditions added for
sub-data.}

\item{s_selection}{is a string value used as datatable option (row datatable
selection).}

\item{l_new_value}{is a list of values associated to the new (sub-data) condtion.
Values are given by the following inputs respectively: "vname", "vtype",
"rel_symbol", "vvalue1"/"vvalue2".}

\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{s_model}{is the model type (3 values: "none", "calib", "valid").}

\item{l_var}{is the list of variables filled in the X, Y, Z and weigthed
residual group selectize inputs.}

\item{s_option}{is a string value (4 values: "label", "color/opacity",
"point type/size", "sorting"). The value is associated to the "edit_option"
selectize input.}

\item{s_sub_option}{is a string value available from the following options
(X/Group qualitative variable): "color/opacity", "point type/size", "sorting".
Group variable is quantitative. \code{s_sub_option} has 2 values ("color", "opacity")
when \code{s_option} = "color/opacity" (related to the "color_opacity" radio button
input). \code{s_sub_option} has 2 values ("type", "size") when \code{s_option} =
"point type/size" (related to the "point_type_size" radio button input).
\code{s_sub_option} has 2 values ("X", "Group") when \code{s_option} = "sorting" (related
to the "x_group_var" radio button input).}

\item{o_option_value}{is a reactive value used to graph label/color/pch/sorting
edition (o_label_text, o_name_option).}

\item{s_plot_type}{is the plot type (4 values: "plot", "boxplot", "histplot",
"barplot").}

\item{b_group}{is a boolean value associated to the group radio button input (T
if a group variable is added, F else).}

\item{i_display}{is a binary value related to the display button status (1:
clicked; 0: not clicked).}

\item{b_pal_auto}{is a boolean value associated to the palette color (T if auto,
F else).}

\item{s_exec}{is a string value used to update label data (2 values: "clear",
"add"). This command is used to clear/add value in label data.}

\item{v_row}{is the vector of row number of label data used to apply the
clear/add command (s_exec).}

\item{s_custom_text}{is a string value added \code{o_option$data} data when the add
command is selected.}

\item{s_dim_num}{is a string value associated to the "dim_num" radio button (2
values: "2d", "3d").}

\item{o_bg_grid}{is a reactive value used to graph background/grid (color)
edition.}

\item{df_all}{is data saved in o_plot reactive value (o_plot$data).}

\item{o_parameter}{is a reactive value used to build the plotly graph.}

\item{s_x_cell_var}{is a variable name corresponding to the selected cell in X
axis (corplot).}

\item{s_y_cell_var}{is a variable name corresponding to the selected cell in Y
axis (corlplot).}

\item{df_xy_cell}{is data associated to a click event on a XY cell of the
correlation matrix (corplot, normal data type).}

\item{i_graph_num}{is an integer value associated to the XY cell graph type (0:
XY scatter-plot + X histplot + Y histplot; 1: XY scatter-plot; 2: X histplot; 3:
Y histplot).}
}
\description{
\code{f_build_modal_dialog_expand} returns a command line used to build a modal
dialog. The modal dialog allows to expand a text area. The text area. The text
area is used to express the (sub-data) condition formula, f(x), g(y), h(z) or the
weighted residuals function.
\cr\code{f_build_modal_dialog_data_option} returns a command line used to build a
modal dialog. The modal dialog allows to specify new options for loaded data
(decimal separator, encoding).
\cr\code{f_create_cond_data} returns a dataset (in datatable format) including all
(sub-data) conditions.
\cr\code{f_check_cond_data} is used to check if a new (sub-data) condition already
exists. An error message (NULL as default) is returned by the function.
\cr\code{f_create_modal_code_var_list} returns a list of two datasets (in datatable
format) including (1) a warning message on missing variable(s) and (2) the
code/variable used to write the function (f(x), g(y), h(z), weighted residuals).
\cr\code{f_create_option_data} allows to create a new dataset (in datatable format)
with customized labels/colors/point characteristics.
\cr\code{f_update_option_data} allows to update the existing dataset (in datatable
format) with customized labels/colors/point characteristics.
\cr\code{f_create_point_type_plotly} returns a plotly used as a point type inventory.
\cr\code{f_create_xy_cell_data} returns data associated to a click event on a XY cell
of the correlation matrix (corplot, normal data type).
\cr\code{f_create_xy_cell_plotly} returns a plotly associated to a click event on a XY
cell of the correlation matrix (corplot, normal data type).
}
