% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_modal_graph.r
\encoding{UTF-8}
\name{f_build_modal_graph}
\alias{f_build_modal_graph}
\title{Modal dialog including a plotly graph used to describe the relationship between
X, Y variables of a correlation matrix.}
\usage{
f_build_modal_graph(df_all, i_num)
}
\arguments{
\item{df_all}{are data created by the \code{f_create_modal_data} function.}

\item{i_num}{is an integer used to build a plotly graph (0: XY scatter-plot + X
histplot + Y histplot; 1: XY scatter-plot; 2: X histplot; 3: Y histplot).}
}
\description{
\code{f_build_modal_graph} allows to create a new graph in a modal dialog when a click
event is executed on a XY cell of the correlation matrix (corplot, normal data
type).
}
