% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_y_axis_range.r
\encoding{UTF-8}
\name{f_calcul_y_axis_range}
\alias{f_calcul_y_axis_range}
\title{Calculation of the Y axis range for the temporal/ir data type}
\usage{
f_calcul_y_axis_range(
  s_data_type = "temporal",
  s_scale = "auto",
  df_all_orig,
  o_parameter,
  o_zoom,
  o_plot,
  o_cond,
  df_click_legend,
  n_frac = 0.05,
  v_range = NULL
)
}
\arguments{
\item{s_data_type}{is the data type (2 values: "temporal", "ir").}

\item{s_scale}{is the value of the "Y scale" radio buttons ("auto" or "manual").}

\item{df_all_orig}{are data saved in the e_data environment ("all" or "sub" if
sub-data is created).}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{o_zoom}{is a reactive value including zoom coordinates of the main plotly
graph.}

\item{o_plot}{is a reactive value including data information corresponding to the
main plotly graph.}

\item{o_cond}{is a reactive value including binary values used as conditions in
the server (related to action buttons).}

\item{df_click_legend}{are legend item data information (name and status).}

\item{n_frac}{is a fraction of the main plotly graph height (value between 0 and
0.1).}

\item{v_range}{is the vector including values filled in "min" and "max" inputs
("Y scale" radio buttons: manual option).}
}
\description{
\code{f_calcul_y_axis_range} is related to the "Y scale" parameter in the Graphic tab
(top panel) and allows to calculate the Y axis range when the auto/manual option
is selected. The function returns a list with the calculated Y range (two ranges
returned for the auto option: range with/without fraction) and a warning message
(if a problem occurs).
}
