% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fields.r
\encoding{UTF-8}
\name{f_check_fields}
\alias{f_check_fields}
\title{Checking process: fields in the section named "Variable selection" (left panel) are
all filled ?}
\usage{
f_check_fields(s_data_type, s_plot_type, s_dim_num, s_model, l_input_id_value)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{s_plot_type}{is the plot type (5 values: "plot", "boxplot", "histplot",
"barplot", "corplot").}

\item{s_dim_num}{is the dimension number (2 values: "2d", "3d").}

\item{s_model}{is the model process (3 values: "none", "calib", "valid").}

\item{l_input_id_value}{is a list of inputs used to display the main plotly graph
with id/value information (id as name).}
}
\description{
\code{f_check_fields} returns a (named) vector of boolean values. Each boolean value is
associated to a graph variable and is true if the field are not filled.
}
