% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.r
\encoding{UTF-8}
\name{f_check_model}
\alias{f_check_model}
\title{Checking process: fields used to calculate the model fitted values are filled in
without error ?}
\usage{
f_check_model(df_all, df_mod_param, o_parameter)
}
\arguments{
\item{df_all}{are data created by the \code{f_prepare_data} function (normal data
type, process 1).}

\item{df_mod_param}{are data saved in the e_data environment (e_data$m_param).}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}
}
\description{
\code{f_check_model} execute a checking process in several steps:
\cr(1) Data matching process applied between values filled in four fields of the
"Variable selection" section (Random, f(x), Weighted residuals: with groups,
variance function) and model parameter data (loaded from the "Model parameter
loading" section);
\cr(2) Checking the variable code in f(x) and variance function fields;
\cr(3) Checking fitted values (vector size, missing/infinite values).
\cr\cr The output is a list including fitted values, transformed Y variable
values (NULL if g(y) not informed), residual variance (NULL if "sigma" not
informed in model paramater data) and an error message if a problem is occured
}
