% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variables.r
\encoding{UTF-8}
\name{f_check_variables}
\alias{f_check_variables}
\title{Checking process: the selected variables are compatible with the data/plot type
requested ?}
\usage{
f_check_variables(
  s_data_type = "normal",
  b_check_plus = F,
  e_data,
  o_parameter
)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{b_check_plus}{is a boolean value such that \code{b_check_plus} is true if the
data type is "normal", the plot type is "plot" and the model process is "none",
false else.}

\item{e_data}{is the environment including loaded data and sub-data (if created).}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}
}
\description{
\code{f_check_variables} execute a checking process in several steps:
\cr(1) Checking fields with the same variables. At least one variable must be
different between fields corresponding to X, Y and Z variables. This process is
not applied for fields corresponding to Random, Weighted residual group and Group
variables when a calibration/validation model is selected;
\cr(2) Check if the ID variable has missing value(s);
\cr(3) Checking the type/number of the selected variables;
\cr(4) Check if the selected variables have no value (combination of all Y
variables for the correlation matrix);
\cr(5) Checking the date variable for the temporal data type (missing values,
date format recognized, unique values);
\cr(6) Check if (".row_num.", ".f.", ".g.", ".h.") variables already exist in
loaded data.
\cr\cr An error message is returned if a problem occured to one of these steps.
}
