% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_click_info.r
\encoding{UTF-8}
\name{f_create_click_info}
\alias{f_create_click_info}
\title{Creating a list of data information associated to a plotly click event}
\usage{
f_create_click_info(
  s_data_type = "ir",
  s_action = "add_flag",
  s_dim_num = NULL,
  v_xy_flag = NULL,
  b_z_flag = F,
  s_flag_name = NA,
  e_data = NULL,
  v_sub_row = NULL,
  o_click_ev,
  o_parameter = NULL,
  o_plot,
  o_cond
)
}
\arguments{
\item{s_data_type}{is the data type (2 values: "normal", "ir").}

\item{s_action}{is the value of the "action" radio button in the Flag tab (2
values: "add_flag", "replace_qc"). This input is only used for the ir data type.}

\item{s_dim_num}{is the dimension number for the normal data type (2 values:
"2d", "3d").}

\item{v_xy_flag}{is a vector of values returned by the "variable" checking boxes
(Flag tab) for ("x", "y") axes.}

\item{b_z_flag}{is a boolean value returned by the "variable" checking boxes
(Flag tab) for the "z" axis.}

\item{s_flag_name}{is the flag data name (saved in the o_flag reactive value:
o_flag$name).}

\item{e_data}{is an environment created in the "WIDEa_launcher" R script. This
input is only used for the normal data type.}

\item{v_sub_row}{is the vector including row numbers of sub-data (NULL as default
value, if no sub-data are created).}

\item{o_click_ev}{are data saved when a click event is executed in the main
plotly graph ("plotly_click" with the \code{event_data} function).}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{o_plot}{is a reactive value including main plotly data information.}

\item{o_cond}{is a reactive value including binary values used as conditions in
the server (related to action buttons).}
}
\description{
\code{f_create_click_info} only concerns the normal/ir data type and allows to save a
list of data information associated to a click event.
\cr\cr For the normal data type, the list includes:
\cr(1) ("x", "y", "z") axes (selected from the "variable" checking boxes in the
Flag tab;
\cr(2) Variable name corresponding to the selected axes;
\cr(3) ID (if an ID variable is selected) or row number from data saved in e_data
environment (e_data$all).
\cr\cr For the ir data type, the list includes:
\cr(1) ID (if an ID variable is selected) or row number from data saved in the
o_plot reactive value (o_plot$id_group);
\cr(2) Row number from data saved in the o_plot reactive value (o_plot$data);
\cr(3) A boolean value returned by the condition "no qc = 1 flags already saved
with this ID or row number;
\cr(4) A warning message returned if a qc = 1 flag is clicked and "add_flag"
value is selected from the "action" radio button in the Flag tab (empty message
else).
}
