% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_concat_variable.r
\encoding{UTF-8}
\name{f_create_concat_variable}
\alias{f_create_concat_variable}
\title{Concatenation of variables}
\usage{
f_create_concat_variable(df_1, v_var)
}
\arguments{
\item{df_1}{are data saved in the e_data environment (e_data$all or e_data$sub).}

\item{v_var}{is the vector of variables.}
}
\description{
\code{f_create_concat_variable} is used to create the concatenated variable (only used
for the normal/ir data type).
}
