% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_current_flag_data.r
\encoding{UTF-8}
\name{f_create_current_flag_data}
\alias{f_create_current_flag_data}
\title{Creating a list including current flag data information}
\usage{
f_create_current_flag_data(
  s_data_type = "normal",
  s_action = NULL,
  s_draw = NULL,
  s_x_var = NULL,
  l_click_info,
  df_previous_flag = NULL,
  e_current_flag,
  o_click_graph = NULL,
  o_plot = NULL
)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{s_action}{is the value of the "action" radio button in the Flag tab (2
values: "add_flag", "replace_qc"). This input is only used for the temporal data
type.}

\item{s_draw}{is the value of the "draw" radio button in the Flag tab (2 values:
"interval", "point"). This input is only used for the temporal data type.}

\item{s_x_var}{is the X variable name. This input is only used for the temporal
data type.}

\item{l_click_info}{is a list of data information associated to a click event and
created by the \code{f_create_click_info} function (only for normal/ir data type). For
the temporal data type, the list returns informations given by the click event
("plotly_click" with the \code{event_data} function).}

\item{df_previous_flag}{are existing flag data given by environments created in
the "WIDEa_server" R script (normal: e_data$flag; temporal:
e_previous_flag$data). This input is only used for the normal/temporal data type.}

\item{e_current_flag}{is the environment created in the "WIDEa_server" R script.}

\item{o_click_graph}{is a reactive value including previous plotly click data
information used to draw an interval for the temporal data type.}

\item{o_plot}{is a reactive value including main plotly data information.}
}
\description{
\code{f_create_current_flag_data} is used to create a list including current flag data
and other ouputs. The number of outputs returned by the list depend on the data
type:
\cr(1) Flag number (normal/temporal);
\cr(2) Data coordinates (normal/ir: 1 dataset; temporal: 2 datasets);
\cr(3) Outputs used to draw an interval (temporal: 2 successive click events
required);
\cr(4) Flag tab button name to be enabled (temporal: interval);
\cr(5) An error message (normal: if the flag already exists; temporal: if the Y
variable is not the same for the 2 successive click events of the interval).
}
