% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_date_variable.r
\encoding{UTF-8}
\name{f_create_date_variable}
\alias{f_create_date_variable}
\title{Creating the date variable}
\usage{
f_create_date_variable(df_all, s_x_var, s_date_format)
}
\arguments{
\item{df_all}{are data saved in the e_data environment (e_data$all).}

\item{s_x_var}{is the X variable name}

\item{s_date_format}{is the date format (values from the format field in
"Variable selection").}
}
\description{
\code{f_create_date_variable} is used to create the date variable (only used for the
temporal data type).
}
