% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_element_data.r
\encoding{UTF-8}
\name{f_create_element_data}
\alias{f_create_element_data}
\title{Creating data related to elements added to the main plotly graph}
\usage{
f_create_element_data(
  s_data_type = "normal",
  i_proc_num = 1,
  o_parameter,
  o_cond,
  o_plot,
  s_mode = "line"
)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{i_proc_num}{is the process number (only used for the temporal/ir data
type).}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{o_cond}{is a reactive value including binary values used as conditions in
the server (related to action buttons).}

\item{o_plot}{is a reactive value including data information corresponding to the
main plotly graph.}

\item{s_mode}{is the value corresponding to the "Mode" parameter in the Graphic
tab (3 values: "line", "marker", "line_marker").}
}
\description{
\code{f_create_element_data} is used to create data including the name (only temporal)
or the number (normal/ir) of elements name added to the main plotly graph when
the display button (left panel) is clicked. This data is used to determine the
position of current flags in the main plotly graph (added with a click event).
The function returns data in different formats: an integer value (normal), a
vector (temporal) and a data frame (ir).
}
