% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flag_data_name.r
\encoding{UTF-8}
\name{f_create_flag_data_name}
\alias{f_create_flag_data_name}
\title{Creating the flag data name}
\usage{
f_create_flag_data_name(s_data_type = "normal", v_split_path, s_id_var = NA)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{v_split_path}{is a vector corresponding to loaded data path splitted by the
"/" symbol.}

\item{s_id_var}{is the ID variable name (NA as default value, if the ID field is
not filled).}
}
\description{
\code{f_create_flag_data_name} used to create the flag data name. The name is composed
of 3 elements: (1) the data name loaded from the "select data" field (left
panel), (2) the data type and (3) the term "withID" if an ID variable is
precised.
}
