% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_data.r
\encoding{UTF-8}
\name{f_create_model_data}
\alias{f_create_model_data}
\title{Creating model data}
\usage{
f_create_model_data(v_f_val, v_vfun_val)
}
\arguments{
\item{v_f_val}{is the fitted values (calculated from the \code{f_check_model}
function).}

\item{v_vfun_val}{is the residual variance (calculated from the \code{f_check_model}
function).}
}
\description{
\code{f_create_model_data} returns a data frame with 1 or 2 columns: (1) fitted
values and (2) residual variance if "sigma" is informed in loaded data
("Model parameter loading" section).
}
