% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_reactive_values.r
\encoding{UTF-8}
\name{f_create_rv_inventory}
\alias{f_create_rv_inventory}
\alias{f_create_rv_id_value_list}
\alias{f_update_rv}
\alias{f_update_graph_option_rv}
\title{List of functions used to update reactive values}
\usage{
f_create_rv_inventory()

f_create_rv_id_value_list(df_rv_id_value, s_choice)

f_update_rv(l_rv_id_value)

f_update_graph_option_rv(
  o_name_option,
  s_option,
  i_display,
  s_data_type,
  s_plot_type,
  v_var_name,
  df_all = NULL,
  b_display_clear = F,
  b_sub_data = F
)
}
\arguments{
\item{df_rv_id_value}{is a data frame of four columns ("choice", "rv", "id",
"value") with reactive value information. The data frame is created from the
\code{f_create_rv_inventory} function.}

\item{s_choice}{is a value used to subset data with reactive value information
(column named "choice" in df_rv_id_value). Five values available: "load1",
"load2", "load3", "display", "lp_s1_s2".}

\item{l_rv_id_value}{is a list with reactive value information (including three
elements "rv", "id", "value").}

\item{o_name_option}{is the reactive value associated to graph option (color,
opacity, point type/size) edition.}

\item{s_option}{is the option value (2 values: "color/opacity",
"point type/size"). \code{s_option} is associated to the "edit_option" selectize input.}

\item{i_display}{is a binary value associated to the display button (1 if
clicked, 0 else).}

\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{s_plot_type}{is the plot type (5 values: "plot", "boxplot", "histplot",
"barplot", "corplot").}

\item{v_var_name}{is the vector of variable name.}

\item{df_all}{are data saved in e_data environment (e_data$all).}

\item{b_display_clear}{is a boolean value associated to the display button.
\code{b_display_clear} is true if the display/clear button is clicked and a graph is
already created, else false.}

\item{b_sub_data}{is a boolean value such that \code{b_sub_data} is true if a sub-data
exists and the "subdata_option" check box input is unchecked, false else.}
}
\description{
\code{f_create_rv_inventory} is used to create inventory of reactive values with id
and initial value. This inventory is used to reset reactive values.
\cr\code{f_create_rv_id_value_list} is used to create a list of id/value used to reset
specific reactive values.
\cr\code{f_update_rv} is used to create command lines to update reactive values.
\cr\code{f_update_graph_option_rv} is used to update reactive values associated to
graph option edition (color/opacity, point type/size).
}
