% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stat_method_message.r
\encoding{UTF-8}
\name{f_create_stat_method_message}
\alias{f_create_stat_method_message}
\title{Data frame including error/warning messages associated to statistic methods added
to the (main) plotly graph}
\usage{
f_create_stat_method_message(
  df_stat_method,
  l_stat_method_level,
  v_stat_method,
  o_parameter,
  df_all
)
}
\arguments{
\item{df_stat_method}{is the statistic method inventory created with the
\code{f_stat_method_ini} function.}

\item{l_stat_method_level}{is a list including group levels (Group selectize
input) for which the checking process has not failed.}

\item{v_stat_method}{is the vector of statistic method name (UI's input ID).}

\item{o_parameter}{is a reactive value including plotly graph information.}

\item{df_all}{are data save in o_plot$data.}
}
\description{
\code{f_create_stat_method_message} returns a data frame including all error/warning
messages for the selected statistic methods (\code{v_stat_method}). An error/warning
message is returned if the checking process associated to statistic methods has
failed (insufficient data, etc.). If a Group variable (selectize input) is added
to the main plotly graph, then a warning message can be returned when the
checking process has not failed for all group levels (else an error message will
be returned).
}
