% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sub_data.r
\encoding{UTF-8}
\name{f_create_sub_data}
\alias{f_create_sub_data}
\title{Creating sub-data}
\usage{
f_create_sub_data(df_all, o_sdata_cond, s_formula)
}
\arguments{
\item{df_all}{are data saved in the e_data environment (e_data$all).}

\item{o_sdata_cond}{is a reactive value including all conditions added.}

\item{s_formula}{is the formula corresponding to the combination of conditions.}
}
\description{
\code{f_create_sub_data} returns a list including (1) row numbers corresponding to the
combination of conditions, (2) the sud-dataset and (3) an error message if a
problem is occured.
}
