% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_axis_layout.r
\encoding{UTF-8}
\name{f_edit_axis_layout}
\alias{f_edit_axis_layout}
\title{Editing the range of the main plotly graph (X, Y, Z) axes or the camera position
for the 3D scatter-plot}
\usage{
f_edit_axis_layout(
  s_data_type = "normal",
  o_parameter,
  o_zoom,
  v_freq_range = NULL
)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{o_zoom}{is a reactive value including zoom coordinates of the main plotly
graph.}

\item{v_freq_range}{is the vector including the spectra frequency range (ir data
type).}
}
\description{
\code{f_edit_axis_layout} returns a list including information on the (X, Y, Z) axes
or the camera position.
}
