% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{f_file_write}
\alias{f_file_write}
\alias{f_numeric_trsf}
\title{List of minor functions used by the WIDEa server}
\usage{
f_file_write(df_flag, s_flag_path)

f_numeric_trsf(n_val, b_pval = F)
}
\arguments{
\item{df_flag}{is the flag data frame.}

\item{s_flag_path}{is the location (path as string value) used to write the flag
data.}

\item{n_val}{is a numeric value.}

\item{b_pval}{is a boolean value equals to TRUE if \code{n_val} corresponds to a
p-value returned by an hypothesis test (FALSE as default value).}
}
\description{
\code{f_file_write} is used to check if it is possible to write flag data. A boolean
value (FALSE) is returned if data are opened.
\cr\code{f_numeric_trsf} is used to round a numeric value and transform it into a
string value.
}
