% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\encoding{UTF-8}
\name{f_load_data}
\alias{f_load_data}
\title{Loading all data}
\usage{
f_load_data(
  s_id,
  s_data_type,
  s_path,
  o_data_opt,
  b_flag = F,
  s_model = "none"
)
}
\arguments{
\item{s_id}{is the action button input ID's used to load data (3 values:
"load1_button", "load2_button", "load3_button").}

\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{s_path}{is the data path.}

\item{o_data_opt}{is a reactive value includig data option: decimal separator,
encoding.}

\item{b_flag}{is a boolean value used to load flag data when the "load1_button"
action button is clicked.}

\item{s_model}{is the model type (3 values: "none", "calib", "valid"). This input
is only used when the "load3_button" action button is clicked.}
}
\description{
\code{f_load_data} returns a list including (1) the loaded data, (2) data information
(related with the load1 button: flag data, flag filename, vector of the code
range for IR data type and a boolean value corresponding to a special enabling of
the "ID" selectize input) and (3) an error/warning message if problem occured
with loaded data.
}
