% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_off_ui_inputs.r
\encoding{UTF-8}
\name{f_on_off_selectize_input}
\alias{f_on_off_selectize_input}
\alias{f_on_off_field_input}
\alias{f_on_off_action_button_input}
\alias{f_on_off_radio_button_input}
\alias{f_on_off_check_box_input}
\alias{f_on_off_check_box_group_input}
\alias{f_init_input_status_data}
\alias{f_create_input_status_list}
\alias{f_update_input_status_list}
\alias{f_update_sp_input_value_list}
\alias{f_update_input_status_data}
\alias{f_on_off_inputs}
\title{List of functions used to enable/disable UI's inputs}
\usage{
f_on_off_selectize_input(
  s_id,
  v_choices,
  b_selected = F,
  i_max_item = NULL,
  i_status,
  i_sp_num = 0,
  i_update = 0,
  b_hide = F
)

f_on_off_field_input(s_id, s_type, i_status, n_value, b_hide = F)

f_on_off_action_button_input(s_id, i_status, s_style = NULL)

f_on_off_radio_button_input(s_id, s_selected, i_status, i_sp_num = 0)

f_on_off_check_box_input(s_id, i_status)

f_on_off_check_box_group_input(s_id, v_selected, i_status)

f_init_input_status_data()

f_create_input_status_list(v_id, v_status)

f_update_input_status_list(l_id_status, df_id_status, b_plot_select = F)

f_update_sp_input_value_list(l_sp_id_values, l_id_status, df_id_status)

f_update_input_status_data(l_id_status, df_id_status, l_sp_id_values = list())

f_on_off_inputs(
  l_id_status,
  df_all = NULL,
  l_selectize_option = list(sub_var_name = NULL, concat = rep(0, 2), max_item = list(),
    sub_data = 0, update = 0),
  l_sp_id_values = list()
)
}
\arguments{
\item{s_id}{is the input ID.}

\item{v_choices}{is the vector of choices for selectize inputs (only used if
\code{i_status} = 1).}

\item{b_selected}{is a boolean value used to add a selected value for selectize
input (if \code{i_status} = 1).}

\item{i_max_item}{is the maximum number of items for a selectize input.}

\item{i_status}{is a binary value corresponding to input status (0: disable ;
1: enable).}

\item{i_sp_num}{is an integer (3 values: 0, 1 and 2) used to execute or not a
special enabling/desabling of UI's inputs. 0 is the default and corresponds to no
special enabling/desabling. 2 is only used for radio buttons.}

\item{i_update}{is a binary value associated to o_cond$update reactive value.}

\item{b_hide}{is a boolean value used to add a command line to hide the input.}

\item{s_type}{is the field input type ("Text" or "Numeric").}

\item{n_value}{is the default value when a numeric input is enabled.}

\item{s_style}{is the action button input style (values: NULL as default,
"default", "danger", "succes").}

\item{s_selected}{is the selected value of radio button input.}

\item{v_selected}{is the vector of selected values for the check box group
inputs.}

\item{v_id}{is the vector of input IDs.}

\item{v_status}{is the vector of input status.}

\item{l_id_status}{is the list of input IDs with status information obtained by
the o_on_off reactive value.}

\item{df_id_status}{is the data frame of input IDs with status information
obtained by the o_input_status reactive value.}

\item{b_plot_select}{is a boolean value related to plot_type and dim_num radio
buttons (T: clicked).}

\item{l_sp_id_values}{is the list of id/value inputs associated to \code{i_sp_num}
value.}

\item{df_all}{is the data frame saved in e_data environment (e_data$all).}

\item{l_selectize_option}{is the selectize input options (five options). The
first option is the variable name of loaded data (e_data$all). The second option
is a vector of binary values given by o_cond reactive value (concat1, concat2).
The third option is a list of inputs with ID and maximum number of items. The
fourth option is a binary value associated to sub-data (= 1 if created). The
creation section).}
}
\description{
\code{f_on_off_selectize_input} is used to create command lines to disable/enable a
selectize input.
\cr\code{f_on_off_field_input} is used to create command lines to disable/enable a
field input (text, numeric).
\cr\code{f_on_off_action_button_input} is used to create command lines to
disable/enable an action button input.
\cr\code{f_on_off_radio_button_input} is used to create command lines to
disable/enable a radio button input.
\cr\code{f_on_off_check_box_input} is used to create command lines to disable/enable a
check box input.
\cr\code{f_on_off_check_box_group_input} is used to create command lines to
disable/enable a check box group input.
\cr\code{f_init_input_status_data} is used to create an input id/status data frame
with initial value. Data are saved in the reactive value named o_input_status.
\cr\code{f_create_input_status_list} is used to create the list of id/status inputs to
be disabled/enabled.
\cr\code{f_update_input_status_list} is used to check the input id/status data frame
(o_input_status) and update the list of inputs to be disabled/enabled. Return a
NULL value if all inputs are already enabled/disabled.
\cr\code{f_update_sp_input_value_list} is used to update the list of id/value special
inputs.
\cr\code{f_update_input_status_data} is used to update the input id/status data frame
(saved in o_input_status).
\cr\code{f_on_off_inputs} is used to aggregate all command lines to disable/enable the
list of specified inputs.
}
