% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\encoding{UTF-8}
\name{f_prepare_data}
\alias{f_prepare_data}
\title{Data preparation for the main plotly graph}
\usage{
f_prepare_data(
  s_data_type = "normal",
  i_proc_num = 1,
  df_all,
  v_sub_row = NULL,
  df_code_freq = NULL,
  v_code_range = NULL,
  df_flag = NULL,
  s_flag_name = NULL,
  df_previous_flag = NULL,
  l_qc = NULL,
  l_id = NULL,
  l_fun_val = NULL,
  o_parameter,
  l_graph_opt = NULL
)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{i_proc_num}{is the process number (only used for the normal data type:
value = 1, 2).}

\item{df_all}{are data saved in the e_data environment (e_data$all or e_data$sub
if sub-data is created).}

\item{v_sub_row}{a vector of row number corresponding to sub-data (only used if
no ID variable is assocciated to flag data.}

\item{df_code_freq}{are data saved in the e_data environment (e_data$code_freq).}

\item{v_code_range}{is the vector of code range saved in the o_data_info
reactive value ("WIDEa_server" R script). This input only concerns the ir data
type.}

\item{df_flag}{are data saved in the e_data environment (e_data$flag).}

\item{s_flag_name}{is the flag data name saved in the o_flag reactive value
("WIDEa_server" R script). This input only concerns the normal data type.}

\item{df_previous_flag}{are data saved in the e_previous_flag environment
(e_previous_flag$data).}

\item{l_qc}{is the list including qc = 1 and 2 data. This input is only used for
the ir data type (data created for the process 1 and required for the process 2).}

\item{l_id}{is the list including id data (id = row number if the ID variable is
not informed). This input is only used for the ir data type (data created for the
process 1 and required for the process 2).}

\item{l_fun_val}{is the list including transformed variable values. This input is
only used for the normal/temporal data type (process 2).}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{l_graph_opt}{is the list of graph option (color, opacity, point type/size)
created by the f_create_graph_opt_vector function. This input is used for the ir
data type (process 2).}
}
\description{
\code{f_prepare_data} is used to prepare data from data (raw/flag) saved in the e_data
environment. All data created from this function are saved in the o_plot
reactive value ("data" for raw data, "data_qc1" and "data_qc2" for flag data with
a qc = 1 or 2 value). The data preparation is performed in two (normal/ir data
type) or three (temporal data type) processes.
\cr(P1.1) Create data (raw and flag by qc value);
\cr(P1.2) Remove qc = 2 flag data in raw data;
\cr(P1.3) Check if the raw data size is superior to 0;
\cr(P1.4) Create a binary vector used to inform if flags (by qc values) exist (ir
data type);
\cr(P1.5) Create data corresponding to isolated points (ir data type);
\cr(P1.6) Check if the range of the variable code in raw data is the same as
code-frequency data (ir data type);
\cr(P1.7) Create data corresponding to isolated points (ir data type).
\cr(P2.1) Replace values of (X, Y, Z) variables by values of variables
transformed by (f, g, h) functions (normal/temporal data type);
\cr(P2.2) Transform data created in the process 1 (transposed data, data
including id/group informations) (ir data type).
\cr(P3) Create data corresponding to isolated points (temporal data type).
}
