% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_current_flag_data.r
\encoding{UTF-8}
\name{f_save_current_flag_data}
\alias{f_save_current_flag_data}
\title{Saving the flag data}
\usage{
f_save_current_flag_data(
  s_data_type = "normal",
  s_action = "add_flag",
  df_previous_flag,
  df_current_flag,
  df_all = NULL,
  s_id_var = NA,
  s_x_var = NA,
  i_qc = NA,
  s_comment = ""
)
}
\arguments{
\item{s_data_type}{is the data type (3 values: "normal", "temporal", "ir").}

\item{s_action}{is the value of "action" radio button in the Flag tab (2 values:
"add_flag", "replace_qc").}

\item{df_previous_flag}{are flag data saved in the e_data environment
(e_data$flag). This input is NULL if the "with flags" box is not checked (left
panel).}

\item{df_current_flag}{are flag data saved in the e_current_flag environment
(e_current_flag$coord).}

\item{df_all}{are data saved in the o_plot reactive value (o_plot$data). This
input is only used for the temporal data type (default = NULL).}

\item{s_id_var}{is the ID variable name (input only used for the normal/IR data
type, default = NA).}

\item{s_x_var}{is the X variable name (input only used for the temporal data
type, default = NA).}

\item{i_qc}{is the value of the "qc" radio button in the Flag tab (input only
used for the temporal/IR data type, default = NA).}

\item{s_comment}{is the comment added from the corresponding field in the Flag
tab (input only used for the temporal/IR data type, default = "").}
}
\description{
\code{f_save_current_flag_data} is used to create/update the flag data. If the
"with flags" box (left panel) is not checked, then a new flag data is created,
else the current flag data (e_current_flag$coord created from the
\code{f_create_current_flag_data} function) is added to the previous (existing) flag
data (e_data$flag).
}
