% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIDEa_server.r
\encoding{UTF-8}
\name{f_server}
\alias{f_server}
\title{Creating the server}
\usage{
f_server(input, output, session)
}
\arguments{
\item{input}{are UI's inputs.}

\item{output}{are UI's outputs.}

\item{session}{is the session object.}
}
\description{
\code{f_server} returns the WIDEa server
}
