% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_method_inventory.r
\encoding{UTF-8}
\name{f_stat_method_ini}
\alias{f_stat_method_ini}
\title{Inventory of statistical methods (Statistics tab) available for all data types}
\usage{
f_stat_method_ini()
}
\description{
\code{f_stat_method_ini} returns a list including (1) a data frame with six columns
(statistic method information) and (2) two lists with information on preliminary
checks executed for statistic methods.
\cr\cr The four columns are:
\cr("data_type") The data type;
\cr("name") The method name;
\cr("check_process") A value used to define which methods have (= 1) or not (= 0)
preliminary checks and which methods have returned an error after preliminary
checks (= -1);
\cr("message") The method name returned by a warning/error message (corresponding
to preliminary checks);
\cr("code") The code associated to preliminary checks applied on statistical
methods. Statistical methods with a same code have same preliminary checks;
\cr("click") A binary value updated when a method is added/removed (check boxes
in Statistics tab) in the current graph.
\cr\cr The two lists return information about (i) group levels (Group selectize
input) added to the main plotly graph legend after preliminary checks have been
applied and (ii) warning/error messages (described by a code: 1, 0 or -1). The
warning/error message is used to inform users about group levels which
preliminary checks failed. 0 (initial value) means no preliminary checks have
been executed for selected statistical methods. The code is equal to 1 (-1 resp.)
if a (no resp.) message must be displayed after preliminary checks.
}
