% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.r
\encoding{UTF-8}
\name{f_ttest}
\alias{f_ttest}
\title{Hypothesis testing}
\usage{
f_ttest(l_lreg, i_coef_num, n_val)
}
\arguments{
\item{l_lreg}{is the linear regression model (\code{lm} class)}

\item{i_coef_num}{is an integer used to select a linear regression coefficient.
By example, if the model is a simple linear regression, then 1 and 2 correspond
to the intercept and the slope respectively.}

\item{n_val}{is the numeric value used by the equality test. By selecting
\code{n_val} = 0, the null hypothesis is to test if the linear regression coefficient
(selected from \code{i_coef_num}) is equal to 0.}
}
\description{
\code{f_ttest} is used to calculate the p-value of an equality test (based on a
Student t distribution) on linear regression (\code{lm} class) coefficients.
}
