% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines.r
\encoding{UTF-8}
\name{f_vline}
\alias{f_vline}
\alias{f_hline}
\alias{f_abline}
\title{Add custom lines to a plotly graph}
\usage{
f_vline(x = 0, color = "black", width = 1, dash = NULL)

f_hline(y = 0, color = "black", width = 1, dash = NULL)

f_abline(
  xmin = 0,
  xmax = 1,
  a = 0,
  b = 1,
  color = "black",
  width = 1,
  dash = NULL
)
}
\arguments{
\item{x}{is the x-axis (numeric) value.}

\item{color}{is a string value corresponding to the line color.}

\item{width}{is a numeric value corresponding to the line width.}

\item{dash}{is a string value equal to "dash" if the line is dashed (NULL as
default).}

\item{y}{is the y-axis (numeric) value.}

\item{xmin}{is the x-axis minimal (numeric) value.}

\item{xmax}{is the x-axis maximal (numeric) value.}

\item{a}{is the intercept (numeric) value.}

\item{b}{is the slope (numeric) value.}
}
\description{
List of funtions used to add vertical (\code{f_vline}), horizontal (\code{f_hline}) and ab
(\code{f_abline}) lines to a plotly graph.
}
