% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIDEa-package.R
\docType{package}
\name{WIDEa-package}
\alias{WIDEa}
\alias{WIDEa-package}
\title{WIDEa: WIDEa: a Web Interface for big Data Exploration, management and analysis}
\description{
WIDEa is R-based software aiming to provide users with a range of functionalities to explore, manage, clean and analyse "big" environmental and (in/ex situ) experimental data. These functionalities are the following, 1. Loading/reading different data types: basic (called normal), temporal, infrared spectra of mid/near region (called IR) with frequency (wavenumber) used as unit (in cm-1); 2. Creation of subsets of data by including several conditions; 3. Interactive data visualization from a multitude of graph outputs (grouped by a qualitative/quantitative variable): 2D/3D scatter-plot, box-plot, hist-plot, bar-plot, correlation matrix; 4. Manipulation of variables: concatenation/re-ordering levels of qualitative variables, transformation of quantitative variables by generic functions in R; 5. Application of statistical methods (linear regression, density curve, hypothesis testing, etc.); 6. Creation/management of data (named flag data) considered as atypical; 7. Study of model results for different strategies: calibration (checking assumptions on residuals), validation (comparison between measured and fitted values). The model considers an error term following a Gaussian distribution and can be written in different forms: (non) linear, mixed effects, main/interaction effects, weighted residuals.
}
\author{
\strong{Maintainer}: Philippe Santenoise \email{philippe.santenoise@inrae.fr}

Other contributors:
\itemize{
  \item  Institut national de recherche pour l'agriculture, l'alimentation et l'environnement (INRAE) [copyright holder]
}

}
\keyword{internal}
