% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lreg.r
\encoding{UTF-8}
\name{f_add_lreg}
\alias{f_add_lreg}
\title{Adding a linear regression in the main plotly graph}
\usage{
f_add_lreg(
  ply_1,
  v_group = "all",
  df_click_legend,
  o_plot,
  o_parameter,
  v_color
)
}
\arguments{
\item{ply_1}{is the plotly object created with the "build_graph" R script.}

\item{v_group}{is the Group variable levels used to applied the corresponding
method. If no Group variable is selected, then \code{v_group} = "all".}

\item{df_click_legend}{are legend item data information (name and status).}

\item{o_plot}{is a reactive value including \code{ply_1} data information.}

\item{o_parameter}{is a reactive value including parameters associated to the
left panel (sections after data loading) and top panels.}

\item{v_color}{is the graph color vector.}
}
\description{
\code{f_add_lreg} returns a plotly object including a linear regression
(\code{lm}). Several quality criteria and hypothesis tests are added in the hover text
of the linear regression: R2, RMSE, equality t-test (\code{f_ttest}: intercept = 0,
slope = 1). This function only concerns 2D plot of the normal data type (disabled
if a calibration model is applied).
}
